Skip rm of not existing repo. -jNUM_CPUS

+++ freebsd/world/build.sh
@@ -47,6 +47,6 @@ jexec ${JAIL_NAME} /usr/bin/make -C /usr/src -j${NUM_CPUS} buildworld buildkerne
 cp -p /etc/resolv.conf ${JAIL_PATH}/etc/
 sed -i .sed.bak s/quarterly/latest/ ${JAIL_PATH}/etc/pkg/FreeBSD.conf
 # clean up old builds
-rm -r ${JAIL_PATH}/usr/obj/usr/src/repo
-jexec ${JAIL_NAME} /usr/bin/make -C /usr/src -j4 packages
+test -d ${JAIL_PATH}/usr/obj/usr/src/repo && rm -r ${JAIL_PATH}/usr/obj/usr/src/repo
+jexec ${JAIL_NAME} /usr/bin/make -C /usr/src -j${NUM_CPUS} packages
 jail -vr ${JAIL_NAME}
